pyproject.toml
[tool.poetry]
name = "font-generation"
version = "0.1.0"
description = ""
authors = ["David Chanin <chanindav@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
fonttools = "^4.25.0"
CairoSVG = "^2.5.2"
Pillow = "^8.3.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = {version = "^21.6b0", allow-prereleases = true}
jupyter = "^1.0.0"
matplotlib = "^3.4.2"
flake8 = "^3.9.2"
mypy = "^0.910"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
target-version = ['py38']
include = '\.pyi?$'
exclude = '''

(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
  | foo.py           # also separately exclude a file named foo.py in
                     # the root of the project
)
'''

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
